fix(#58): move DTO to dtos/, fix CHANGELOG typo
